+++ mountpoint-s3-crt/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update to latest CRT dependencies
 * Default exponential backoff value retries is now 500ms rather than 25ms ([awslabs/aws-c-io#612](https://github.com/awslabs/aws-c-io/pull/612))
+* Renamed `s3::RequestType::Default` to `s3::RequestType::Unknown`
 
 ## v0.4.0 (November 21, 2023)